requirements.txt
Jinja2==3.1.3
pytest==8.0.1
PyYAML==6.0.1
PyYAML==6.0.1
pyyaml_include==1.3.2
setuptools==69.1.0
